callautomation-appointment-reminder/requirements.txt
aiohttp==3.7.4.post0
azure-communication-callautomation @file:///C:/pyenv2_311/azure_communication_callautomation-1.0.0a20230413001-py3-none-any.whl
azure-communication-identity==1.3.1